requirements.txt
streamlit 
pillow 
pytesseract 
transformers
ollama
easyocr
groq
opencv-python
